% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw.R
\name{clean_raw}
\alias{clean_raw}
\title{Clean extracted datasets}
\usage{
clean_raw(indir, outdir, obs = 0.95, hole = 3, tocc = 0.3, len = 5)
}
\arguments{
\item{indir}{The directory where the extracted datasets are located}

\item{outdir}{The directory to save the preprocessed datasets}

\item{obs}{Minimum acceptable observation rate}

\item{hole}{Minimum acceptable hole size in minutes}

\item{tocc}{Occupancy threshold}

\item{len}{Minimum length of a congested period in minutes}
}
\value{
Preprocessed datasets in csv format
}
\description{
This function is applied to filter datasets with low observation rate, long missing gap, impute missing values, and filter datasets with insufficient congestion periods
}
