% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_filtered.R
\name{cut_filtered}
\alias{cut_filtered}
\title{Cut filtered datasets}
\usage{
cut_filtered(inDir, outDir, remove)
}
\arguments{
\item{inDir}{The directory where filtered datasets are stored}

\item{outDir}{The directory where cut datasets are stored}

\item{remove}{A vector of lane numbers that you want to remove}
}
\value{
Cut datasets in the format of csv
}
\description{
The function is to select a subset of lane traffic flow characteristics from each dataset
}
