% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_stepwise.R
\name{do_stepwise}
\alias{do_stepwise}
\title{Stepwise duration-occupancy relation}
\usage{
do_stepwise(occ, not, hot, xot, ndr, hdr, xdr)
}
\arguments{
\item{occ}{Candidate occupancy, which is the input variable}

\item{not}{Normal-occupancy threshold}

\item{hot}{High-occupancy threshold(*)}

\item{xot}{Extremely high-occupancy threshold}

\item{ndr}{The minimum acceptable duration in normal-occupancy periods(*)}

\item{hdr}{The minimum acceptable duration in high-occupancy periods(*)}

\item{xdr}{The minimum acceptable duration in extremely high-occupancy periods}
}
\value{
Minimum acceptable duration
}
\description{
This function is applied to calculate the minimum acceptable duration given the occupancy based on a stepwise duration-occupancy relation
}
