% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_unitgap.R
\name{do_unitgap}
\alias{do_unitgap}
\title{Unit gap duration-occupancy relation}
\usage{
do_unitgap(occ, lot, hot, xhot, ldr, hdr, xhdr)
}
\arguments{
\item{occ}{Candidate occupancy, which is the input variable}

\item{lot}{Low-occupancy threshold(*)}

\item{hot}{High-occupancy threshold}

\item{xhot}{Extremely high-occupancy threshold(*)}

\item{ldr}{The minimum acceptable duration in low-occupancy periods(*)}

\item{hdr}{The minimum acceptable duration in high-occupancy periods}

\item{xhdr}{The minimum acceptable duration in extremely high-occupancy periods(*)}
}
\value{
Minimum acceptable duration
}
\description{
This function is applied to calculate the minimum acceptable duration given the occupancy based on an inverse linear duration-occupancy relation
}
