% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw.R
\name{extract_raw}
\alias{extract_raw}
\title{Extract raw datasets from PeMS}
\usage{
extract_raw(indir, outdir, id, lanes)
}
\arguments{
\item{indir}{The directory where the PeMS raw gz files are located}

\item{outdir}{The directory to save the extracted csv files}

\item{id}{station ID, which is the VDS in PeMS}

\item{lanes}{number of lanes}
}
\value{
Extracted raw datasets in csv format
}
\description{
This function is applied to extract raw datasets from Data Clearinghouse of PeMS and save the extracted datasets in csv format.
}
