% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate.R
\name{get_candidate}
\alias{get_candidate}
\title{Get candidate intervals}
\usage{
get_candidate(datasets, penalty)
}
\arguments{
\item{datasets}{The normalized datasets in list format}

\item{penalty}{The penalty for the PELT changepoint detection method}
}
\value{
List of candidate intervals
}
\description{
This function is applied to detect candidate intervals based on the PELT changepoint detection method
}
