% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized.R
\name{get_normalized}
\alias{get_normalized}
\title{Get normalized datasets}
\usage{
get_normalized(datasets)
}
\arguments{
\item{datasets}{A list of filtered datasets}
}
\value{
A list of normalized datasets
}
\description{
The function is to normalize total vehicle c
ounts and occupancies
}
