% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stationary.R
\name{get_stationary}
\alias{get_stationary}
\title{Get near-stationary states}
\usage{
get_stationary(datasets, mindr = 4, maxdr = 10, hot = 0.4, gapt)
}
\arguments{
\item{datasets}{The candidate intervals in list format}

\item{mindr}{The minimum acceptable duration of free-flow period}

\item{maxdr}{The minimum acceptable duration of congested period}

\item{hot}{High occupancy threshold}

\item{gapt}{Gap threshold}
}
\value{
List of near-stationary states
}
\description{
This function is applied to identify near-stationary states from candidate intervals based on the modified Cassidy's method
}
