% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistics.R
\name{get_statistics}
\alias{get_statistics}
\title{Measure statistics of NSS}
\usage{
get_statistics(datasets, critical, tolerance)
}
\arguments{
\item{datasets}{The datasets of identified near-stationary states in the list format.}

\item{critical}{Pre-determined critical density by prior knowledge, e.g. 0.18}

\item{tolerance}{Minimum acceptable sample size; R-squared would be regarded zero if the sample size is smaller than the tolerance.}
}
\description{
This function is applied to measure quantity and quality of identified near-stationary states in the congested regime.
}
