% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filtered.R
\name{read_filtered}
\alias{read_filtered}
\title{Read preprocessed data}
\usage{
read_filtered(dir)
}
\arguments{
\item{dir}{The directory of the preprocessed datasets}
}
\value{
List of preprocessed datasets
}
\description{
This function is applied to read preprocessed datasets in a list
}
