% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_normalized.R
\name{get_normalized}
\alias{get_normalized}
\title{Get normalized series}
\usage{
get_normalized(datasets, rem = NULL)
}
\arguments{
\item{datasets}{The preprocessed list}

\item{rem}{Remove any lanes when calculating averaged count and occupancy series?}
}
\value{
List of normalized datasets
}
\description{
This function is applied to calculate normalized count and occupancy series
}
